Turn off ale_echo_cursor
On further investigation, it's the echo cursor feature that moves the cursor one left in insert-normal mode at the end of a line. That only happens when there's an error or warning on that line. So I'm turning that off and reenabling ale_lint_on_text_changed. Instead of relying on echo cursor, we can use :lopen to see the error messages.